// src/app/model/Publicacion.model.ts
export class Publicacion {
    id_publicacion: number | undefined;
    nombre: string | undefined;
    imagen: string | undefined;
    fecha_publicacion: Date | undefined;
    autor: string | undefined;
    descripcion: string | undefined;
    id_hemeroteca: number | undefined;
    id_tipo: number | undefined;
    categoria: string | undefined;
    nombre_hemeroteca: string | undefined;
}